requirements.txt
pytest>=7.0
python-dotenv>=1.0
openai>=1.0
chromadb>=0.4.0,<1.0
